+++ timeutil/timeutil_test.go
@@ -160,7 +160,9 @@ func (s *TimeUtilSuite) TestTinyDate(c *C) {
 
 	c.Assert(td.Unix(), Equals, dt.Unix())
 	c.Assert(td.Time().Unix(), Equals, dt.Unix())
-	c.Assert(TinyDate(1000), Equals, Date(0))
+	c.Assert(td.IsZero(), Equals, false)
+
+	c.Assert(TinyDate(1000).IsZero(), Equals, true)
 }
 
 func (s *TimeUtilSuite) TestDateNames(c *C) {